ml_service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
pdfplumber==0.11.8
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
gitpython==3.1.40
pymongo==4.6.1
psycopg2-binary==2.9.9
redis==5.0.1
prophet==1.1.5
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
